codice-civile-ai/requirements.txt
fastapi
uvicorn
pdfplumber
chromadb
sentence-transformers
langchain
langchain-community
langchain-anthropic
langchain-openai
anthropic
openai
python-dotenv
pydantic
slowapi
